Check UID's process state before restricting its network.
It is possible that an app which is in fg-service process state to be
considered idle. Since we don't want to block connectivity of apps in
the foreground state, we should check the uid process state in addition
to its idle state when deciding whether to block its network or not.

Test: atest com.android.cts.net.HostsideRestrictBackgroundNetworkTests#testAppIdleMetered_enabled

Signed-off-by: xiamengsen <xiamengsen@xiaomi.com>
Change-Id: If7e42605eb162abf6f88f03f28f7245659a7748c
Signed-off-by: starlight5234 <starlight5234@protonmail.ch>